+++ frontend/src/pages/payment_icon_yellow_large.png
